# Miscellaneous/openapi.yaml
openapi: 3.0.0
info:
  title: TourPlanner
  description: This is the specification of the API endpoints of the tourplanner server.
  version: 1.0.0
...
paths:
...
  /import:
    post:
      tags:
        - Data transfer
      summary: Import tours
      description: Import tours
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tour'
      responses:
        '201':
          description: Tours imported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tour'
        '204':
          description: The request was fine, but no tours were created yet
        '500':
          description: Internal server error
  /export:
    get:
      tags:
        - Data transfer
      summary: Export tours
      description: Export tours
      responses:
        '200':
          description: A minimum of one tour exists, the response contains an array of tours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tour'
        '204':
          description: The request was fine, but no tours were created yet
        '500':
          description: Internal server error
...
components:
  schemas:
...
    Tour:
      type: object
      properties:
        tourId:
          type: integer
          example: 1
        name:
          type: string
          example: "MyTour"
        tourDescription:
          type: string
          example: "My beautiful tour"
        start:
          type: string
          example: "Vienna"
        destination:
          type: string
          example: "Bratislava"
        transportType:
          type: string
          example: "car"
          enum:
            - "car"
            - "bicycle"
            - "pedestrian"
        tourDistance:
          type: number
          example: 80.5
          description: Tour distance in kilometers
        estimatedTime:
          type: integer
          example: 3480000
          description: Estimated time in milliseconds
        tourInformation:
          type: string
          example: aHR0cHM6Ly93d3cueW91dHViZS5jb20vd2F0Y2g/dj1kUXc0dzlXZ1hjUQ==
          description: Base64 encoded image file
        tourLogs:
          type: array
          items:
            $ref: '#/components/schemas/TourLog'
    TourLog:
      type: object
      properties:
        tourLogId:
          type: integer
          example: 1
        date:
          type: string
          example: "2023-05-11 10:16:51"
          description: Date YYYY-MM-DD HH:MM:SS
        comment:
          type: string
          example: "Me likes!"
        difficulty:
          type: integer
          example: 2
          description: Difficulty (from 1 to 5)
        totalTime:
          type: integer
          example: 3480000
          description: Actual time spent in milliseconds
        rating:
          type: integer
          example: 5
          description: Rating (1-5 stars)